+++ lisp/eab-org-publish.el
@@ -3,7 +3,7 @@
 	       ("html-fz-base")))
 (add-to-list 'org-publish-project-alist
 	     `("html-fz-base"
-	       :base-directory "/home/eab/git/fz/tasks/sync-jira/fz-wiki/org"
+	       :base-directory "/home/eab/git/fz/db/sync-jira/fz-wiki/org"
 	       :publishing-directory "/home/eab/git/eabmisc/boykov.github.io/fz-infra"
 	       :online-suffix ".html"
 	       :working-suffix ".org"
@@ -21,7 +21,7 @@
 		"html-ldap-static")))
 (add-to-list 'org-publish-project-alist
 	     `("html-ldap-base"
-	       :base-directory "/home/eab/git/cc/cc-ldap-centos/docs"
+	       :base-directory "/home/eab/git/its/cc-ldap-centos/docs"
 	       :publishing-directory "/home/eab/git/eabmisc/boykov.github.io/ldap"
 	       :online-suffix ".html"
 	       :working-suffix ".org"
@@ -34,7 +34,7 @@
 	       :creator-info nil))
 (add-to-list 'org-publish-project-alist
 	     `("html-ldap-static"
-	       :base-directory "/home/eab/git/cc/cc-ldap-centos/docs/"
+	       :base-directory "/home/eab/git/its/cc-ldap-centos/docs/"
 	       :base-extension "png"
 	       :publishing-directory "/home/eab/git/eabmisc/boykov.github.io/ldap"
 	       :recursive t
@@ -80,7 +80,16 @@
     (shell-command
      (concat eab/ssh-host " <<'END'
 docker exec eab-utils bash -c \"cd ~/git/eab-kb/js && node update-client.js\"
-END"))))
+END"))
+    (if (equal
+	 (file-truename (buffer-file-name (buffer-base-buffer)))
+	 "/home/eab/pnt/jaguar/git/org-chronos/clock/w1c-plan-vacancy.org")
+	(progn
+	  (org-babel-tangle)
+	  (shell-command
+	   (concat "ssh chronos <<'END'
+screen -d -m bash -c \"cd ~/git/eabmisc/eablatex && make resume\"
+END"))))))
 
 (setq org-html-head-include-default-style nil)